new structure for effect classes, in order to avoid duplication
