+++ alchemist/alchemist-swingui/src/main/java/it/unibo/alchemist/boundary/gui/effects/DrawLayersGradient.java
@@ -11,7 +11,6 @@ package it.unibo.alchemist.boundary.gui.effects;
 
 import it.unibo.alchemist.boundary.wormhole.interfaces.IWormhole2D;
 import it.unibo.alchemist.model.interfaces.Environment;
-import it.unibo.alchemist.model.interfaces.Layer;
 import it.unibo.alchemist.model.interfaces.Position2D;
 import org.danilopianini.lang.RangedInteger;
 import org.danilopianini.view.ExportForGUI;
@@ -20,7 +19,6 @@ import java.awt.Point;
 import java.awt.Graphics2D;
 import java.awt.Color;
 import java.awt.geom.Dimension2D;
-import java.util.Collection;
 import java.util.function.Function;
 
 /**
@@ -38,7 +36,6 @@ import java.util.function.Function;
  */
 public abstract class DrawLayersGradient extends DrawLayersValues {
 
-    private static final long serialVersionUID = 1L;
     private static final int MIN_SAMPLES = 10;
     private static final int MAX_SAMPLES = 400;
 
@@ -49,46 +46,57 @@ public abstract class DrawLayersGradient extends DrawLayersValues {
      * {@inheritDoc}
      */
     @Override
-    protected abstract <T, P extends Position2D<P>> void drawLayers(Collection<Layer<T, P>> toDraw, Environment<T, P> env, Graphics2D g, IWormhole2D<P> wormhole);
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected <T, P extends Position2D<P>> void drawValues(final Function<? super P, ? extends Number> f, final Environment<T, P> env, final Graphics2D g, final IWormhole2D<P> wormhole) {
+    public <T, P extends Position2D<P>> void drawFunction(final Function<? super P, ? extends Number> f, final Environment<T, P> env, final Graphics2D g, final IWormhole2D<P> wormhole) {
         if (minOrMaxLayerValuesNeedsToBeUpdated()) {
             updateMinAndMaxLayerValues();
         }
+        // to draw the gradient, we simply divide the screen into cells, then we
+        // visit each cell and determine its value by interpolating the values
+        // at the corners. We then map such value to a color and fill the cell
+        // with that color
         final Dimension2D viewSize = wormhole.getViewSize();
         final int viewStartX = 0;
         final int viewStartY = 0;
         final int viewEndX = (int) Math.ceil(viewSize.getWidth());
         final int viewEndY = (int) Math.ceil(viewSize.getHeight());
+        // step is the side of each cell
         final int stepX = (viewEndX - viewStartX) / samples.getVal();
         final int stepY = (viewEndY - viewStartY) / samples.getVal();
+        // visiting a cell means having the indexes of its four corners
         for (int i1 = viewStartX; i1 < viewEndX; i1 += stepX) {
             final int i2 = i1 + stepX;
             for (int j1 = viewStartY; j1 < viewEndY; j1 += stepY) {
                 final int j2 = j1 + stepY;
+                // the four points of the view
                 final Point p1 = new Point(i1, j1);
                 final Point p2 = new Point(i1, j2);
                 final Point p3 = new Point(i2, j1);
                 final Point p4 = new Point(i2, j2);
+                // we map them to env points
                 final P envP1 = wormhole.getEnvPoint(p1);
                 final P envP2 = wormhole.getEnvPoint(p2);
                 final P envP3 = wormhole.getEnvPoint(p3);
                 final P envP4 = wormhole.getEnvPoint(p4);
+                // get the values
                 final double v1 = f.apply(envP1).doubleValue();
                 final double v2 = f.apply(envP2).doubleValue();
                 final double v3 = f.apply(envP3).doubleValue();
                 final double v4 = f.apply(envP4).doubleValue();
+                // interpolate such values
                 final double v = (v1 + v2 + v3 + v4) / 4;
+                // fill the cell with the color
                 final double newAlpha = map(v, getMinLayerValueDouble(), getMaxLayerValueDouble(), 0, getAlpha().getVal());
                 g.setColor(new Color(getRed().getVal(), getGreen().getVal(), getBlue().getVal(), (int) Math.ceil(newAlpha)));
                 g.fillRect(i1, j1, i2 - i1, j2 - j1);
             }
         }
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected abstract LayerToFunctionMapper createMapper();
 
     /**
      * @return the number of samples
