+++ alchemist/alchemist-swingui/src/main/kotlin/it/unibo/alchemist/boundary/gui/effects/BidimensionalGaussianLayersMapper.kt
@@ -1,0 +1,57 @@
+/*
+ * Copyright (C) 2010-2019, Danilo Pianini and contributors
+ * listed in the main project's alchemist/build.gradle.kts file.
+ *
+ * This file is part of Alchemist, and is distributed under the terms of the
+ * GNU General Public License, with a linking exception,
+ * as described in the file LICENSE in the Alchemist distribution's top directory.
+ */
+
+package it.unibo.alchemist.boundary.gui.effects
+
+import it.unibo.alchemist.boundary.wormhole.interfaces.IWormhole2D
+import it.unibo.alchemist.model.implementations.layers.BidimensionalGaussianLayer
+import it.unibo.alchemist.model.interfaces.Environment
+import it.unibo.alchemist.model.interfaces.Layer
+import it.unibo.alchemist.model.interfaces.Position2D
+import java.awt.Graphics2D
+import java.util.function.Function
+import java.util.stream.Collectors
+import java.util.stream.Stream
+
+/**
+ * Maps [BidimensionalGaussianLayer]s, it ignores any other layer.
+ *
+ * This class also manages to infer optimal min and max layer values automatically
+ * so the user does not have to set them by hand.
+ */
+class BidimensionalGaussianLayersMapper : LayerToFunctionMapper {
+
+    private var minAndMaxToBeSet = true
+
+    override fun <T, P : Position2D<P>> prepare(effect: DrawLayersValues, toDraw: Collection<Layer<T, P>>, env: Environment<T, P>, g: Graphics2D, wormhole: IWormhole2D<P>) {
+        if (minAndMaxToBeSet) {
+            val minLayerValue = 0.1
+            val maxLayerValue = toDraw.stream()
+                .filter { l -> l is BidimensionalGaussianLayer }
+                .map { l -> l as BidimensionalGaussianLayer }
+                .map { l -> l.getValue(env.makePosition(l.centerX, l.centerY)) }
+                .max { d1, d2 -> java.lang.Double.compare(d1, d2) }
+                .orElse(minLayerValue)
+            effect.minLayerValue = minLayerValue.toString()
+            effect.maxLayerValue = maxLayerValue.toString()
+            minAndMaxToBeSet = false
+        }
+    }
+
+    override fun <T, P : Position2D<P>> map(layers: Collection<Layer<T, P>>): Collection<Function<in P, out Number>> {
+        return map(layers.stream()).collect(Collectors.toList())
+    }
+
+    override fun <T, P : Position2D<P>> map(layers: Stream<Layer<T, P>>): Stream<Function<in P, out Number>> {
+        return layers
+            .filter { l -> l is BidimensionalGaussianLayer }
+            .map { l -> l as BidimensionalGaussianLayer }
+            .map { l -> Function { p: P -> l.getValue(p) } } // l::getValue
+    }
+}
